Port emulation of The Voice
Standalone O2EM can emulate The Voice expansion module using wav sound
samples that the user must place in a specific folder, similarly to
what happens with the bios. The libretro port of O2EM, however, didn't
include The Voice support, as the corresponding code was simply
disabled.

This commit ports The Voice emulation to work in a libretro context,
using libretro-common's own mixer to load and play the sound samples,
which should be placed in a 'voice' folder under the usual 'System'
folder.