Expose external RAM through retro_get_memory_data
The core only exposed the internal RAM, but many games make extensive
use of the external RAM as well. Having no access to it is a problem
for developing achievements.

External RAM is now exposed through `RETRO_MEMORY_SYSTEM_RAM`, right
after the internal RAM. This memory map is virtual, as the console
itself accesses the two RAMs using specific instructions instead of
mapping them to a single address space.